Removed jest main.e2e.spec.ts and minor changes on PW main.e2e.spec.ts
