+++ e2e/main.e2e.spec.ts
@@ -1,95 +1,110 @@
+import path from 'path';
 import os from 'os';
-import { Application, SpectronClient } from 'spectron';
-import { BrowserWindow } from 'electron';
-import NavBarPage from './pages/navbar';
-import { TestUtils } from './utils/TestUtils';
+import { ElectronApplication, _electron, Page, Locator } from 'playwright';
+import { test, expect } from '@playwright/test';
 
-describe('Rancher Desktop', () => {
-  let utils: TestUtils;
-  let app: Application;
-  let client: SpectronClient;
-  let browserWindow: BrowserWindow;
-  let navBarPage: NavBarPage;
+let page: Page;
+let electronApp: ElectronApplication;
 
-  beforeAll(async() => {
-    utils = new TestUtils();
-    utils.setupJestTimeout();
-    app = await utils.setUp();
+/**
+ * Using test.describe.serial make the test execute step by step, as described on each `test()` order
+ * Playwright executes test in parallel by default and it will not work for our app backend loading process.
+ * */
+test.describe.serial('Rancher Desktop - Main App', () => {
+  let mainTitle: Locator;
+  const mainTitleSelector = '[data-test="mainTitle"]';
 
-    if (app && app.isRunning()) {
-      client = app.client;
-      browserWindow = app.browserWindow;
-      navBarPage = new NavBarPage(app);
-
-      return utils.byPassFirstPage();
-    } else {
-      console.log('Application error: Does not started properly');
-    }
+  test.beforeAll(async() => {
+    electronApp = await _electron.launch({ args: [path.join(__dirname, '../')] });
+    const appPath = await electronApp.evaluate(async({ app }) => {
+      return await app.getAppPath();
+    });
   });
 
-  afterAll(async(done) => {
-    await utils.tearDown();
-    done();
+  test.afterAll(async() => {
+    await electronApp.close();
   });
 
-  it('opens the window', async() => {
-    await client.waitUntilWindowLoaded();
-    const title = await browserWindow.getTitle();
-
-    expect(title).toBe('Rancher Desktop');
+  test('should open the main app', async() => {
+    page = await electronApp.firstWindow();
+    await page.waitForSelector('.progress');
   });
 
-  it('should display welcome message in General tab', async() => {
-    const generalPage = await navBarPage.getGeneralPage();
+  test('should land on General page', async() => {
+    mainTitle = page.locator(mainTitleSelector);
 
-    expect(generalPage).not.toBeNull();
-    expect(await generalPage?.getMainTitle()).toBe('Welcome to Rancher Desktop');
+    await await expect(mainTitle).toHaveText('Welcome to Rancher Desktop');
   });
 
-  it('should switch to Kubernetes Settings tab', async() => {
-    const kubernetesPage = await navBarPage.getKubernetesPage();
+  test('should navigate to Kubernetes Settings and check elements', async() => {
+    const k8sMemorySliderSelector = '[id="memoryInGBWrapper"]';
+    const k8sCpuSliderSelector = '[id="numCPUWrapper"]';
+    const k8sPortSelector = '[data-test="portConfig"]';
+    const k8sResetBtn = '[data-test="k8sResetBtn"]';
 
-    expect(kubernetesPage).not.toBeNull();
-    expect(await kubernetesPage?.getK8sVersionDropDown()).toBeTruthy();
+    await navigateTo('K8s');
+    // Collecting data from selectors
+    const k8sSettingsTitle = page.locator(mainTitleSelector);
+    const k8sMemorySlider = page.locator(k8sMemorySliderSelector);
+    const k8sCpuSlider = page.locator(k8sCpuSliderSelector);
+    const k8sPort = page.locator(k8sPortSelector);
+    const k8sResetButton = page.locator(k8sResetBtn);
 
-    if (os.platform().startsWith('win')) {
-      expect(await kubernetesPage?.getK8sCpuConfig()).toBeFalsy();
-      expect(await kubernetesPage?.getK8sMemoryConfig()).toBeFalsy();
-    } else {
-      expect(await kubernetesPage?.getK8sMemoryConfig()).toBeTruthy();
-      expect(await kubernetesPage?.getK8sCpuConfig()).toBeTruthy();
-    }
-
-    expect(await kubernetesPage?.getK8sPortConfig()).toBeTruthy();
-    expect(await kubernetesPage?.getMainTitle()).toBe('Kubernetes Settings');
-    expect(await kubernetesPage?.getResetKubernetesButtonText()).toBe('Reset Kubernetes');
+    await expect(k8sSettingsTitle).toHaveText('Kubernetes Settings');
+    await expect(k8sMemorySlider).toBeVisible();
+    await expect(k8sCpuSlider).toBeVisible();
+    await expect(k8sPort).toBeVisible();
+    await expect(k8sResetButton).toBeVisible();
   });
 
-  it('should switch to Port Forwarding tab', async() => {
-    const portForwardingPage = await navBarPage.getPortForwardingPage();
+  if (os.platform().startsWith('win')) {
+    test('should navigate to Port Forwarding anc check elements', async() => {});
+  } else {
+    test('should navigate to Supporting Utilities anc check elements', async() => {
+      await navigateTo('Integrations');
+      const getSupportTitle = page.locator(mainTitleSelector);
 
-    if (os.platform().startsWith('win')) {
-      expect(portForwardingPage).not.toBeNull();
-    } else {
-      expect(portForwardingPage).toBeNull();
-    }
+      await expect(getSupportTitle).toHaveText('Supporting Utilities');
+    });
+  }
+
+  test('should navigate to Images page', async() => {
+    const getSupportTitle = page.locator(mainTitleSelector);
+
+    await navigateTo('Images');
+    await expect(getSupportTitle).toHaveText('Images');
   });
 
-  it('should switch to Images tab', async() => {
-    const imagesPage = await navBarPage.getImagesPage();
+  test('should navigate to Troubleshooting and check elements', async() => {
+    const getSupportTitle = page.locator(mainTitleSelector);
 
-    expect(imagesPage).not.toBeNull();
-    expect(await imagesPage?.getImagesTable()).toBeTruthy();
-    expect(await imagesPage?.getMainTitle()).toBe('Images');
-  });
-
-  it('should switch to Troubleshooting tab', async() => {
-    const troubleShootingPage = await navBarPage.getTroubleshootingPage();
-
-    expect(troubleShootingPage).not.toBeNull();
-    expect(await troubleShootingPage?.getTroubleshootingDashboard()).toBeTruthy();
-    expect(await troubleShootingPage?.getLogsButton()).toBeTruthy();
-    expect(await troubleShootingPage?.getFactoryResetButton()).toBeTruthy();
-    expect(await troubleShootingPage?.getMainTitle()).toBe('Troubleshooting');
+    await navigateTo('Troubleshooting');
+    await expect(getSupportTitle).toHaveText('Troubleshooting');
   });
 });
+
+/**
+ * Navigate to a specific tab
+ * @param path
+ */
+async function navigateTo(path: string) {
+  try {
+    return await Promise.all([
+      page.click(`.nav li[item="/${ path }"] a`),
+      page.waitForNavigation({ url: `**/${ path }` })
+    ]);
+  } catch (err) {
+    console.log(`Cannot navigate to ${ path }. Error ---> `, err);
+  }
+}
+
+/**
+ * Delay function to slow things down
+ * @param time
+ * @returns
+ */
+function delay(time: number | undefined) {
+  return new Promise((resolve) => {
+    setTimeout(resolve, time);
+  });
+}
